feat: whatsapp.mjs > ts
