+++ src/whatsapp/index.ts
@@ -1,0 +1,2 @@
+export * from './whatsapp';
+export * from './interfaces';
