+++ src/whatsapp/interfaces/message.ts
@@ -1,0 +1,4 @@
+export interface MessageApp {
+   text: string;
+   contact: string;
+}
